pyproject.toml
[project]
name = "crag-reflective-rag"
version = "1.0.0"
description = "Corrective RAG + Self-Reflective RAG Implementation"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["rag", "crag", "self-reflective", "fastapi", "qdrant", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.17",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "docling>=2.11.0",
    "docling-core>=2.5.0",
    "qdrant-client>=1.12.0",
    "openai>=1.54.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "numpy>=2.1.0",
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
crag-rag = "app.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
